react-agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "chardet>=5.2.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.33",
    "langchain-text-splitters>=0.3.11",
    "langgraph>=0.6.7",
    "langsmith>=0.4.31",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pypdf>=6.1.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
